requirements.txt
aioredis==2.0.1
alembic==1.11.3
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.102.0
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.2
kombu==5.3.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pendulum==2.1.2
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
